Update XDMA to add the accompanied cfg
