+++ hw/chisel/src/main/scala/snax/xdma/xdmaFrontend/XDMADataPath.scala
@@ -17,39 +17,11 @@ import snax.readerWriter.{
 }
 import snax.xdma.DesignParams._
 import snax.DataPathExtension._
-
-// The ReaderWriterCfg Class that used for interface between local Datapath and DMA Ctrl
-// The length of addresses is the short version, which is just enough to reside TCDM
-class XDMADataPathCfgIO(param: XDMADataPathParam) extends Bundle {
-  val aguCfg =
-    new AddressGenUnitCfgIO(param =
-      param.rwParam.aguParam
-    ) // Buffered within AGU
-  val readerwriterCfg = new ReaderWriterCfgIO(param.rwParam)
-  val loopBack = Bool()
-
-  val extCfg = if (param.extParam.length != 0) {
-    Vec(
-      param.extParam.map { i => i.extensionParam.userCsrNum }.reduce(_ + _) + 1,
-      UInt(32.W)
-    ) // The total csr required by all extension + 1 for the bypass signal
-  } else Vec(0, UInt(32.W))
-
-  def connectWithList(
-      csrList: IndexedSeq[UInt]
-  ): IndexedSeq[UInt] = {
-    var remaincsrList = csrList
-    remaincsrList = aguCfg.connectWithList(remaincsrList)
-    remaincsrList = readerwriterCfg.connectWithList(remaincsrList)
-    extCfg := remaincsrList.take(extCfg.length)
-    remaincsrList = remaincsrList.drop(extCfg.length)
-    remaincsrList
-  }
-}
+import snax.xdma.io._
 
 class XDMADataPath(
-    readerparam: XDMADataPathParam,
-    writerparam: XDMADataPathParam,
+    readerParam: XDMAParam,
+    writerParam: XDMAParam,
     clusterName: String = "unnamed_cluster"
 ) extends Module
     with RequireAsyncReset {
@@ -58,8 +30,8 @@ class XDMADataPath(
 
   val io = IO(new Bundle {
     // All config signal for reader and writer
-    val readerCfg = Input(new XDMADataPathCfgIO(readerparam))
-    val writerCfg = Input(new XDMADataPathCfgIO(writerparam))
+    val readerCfg = Input(new XDMACfgIO(readerParam))
+    val writerCfg = Input(new XDMACfgIO(writerParam))
 
     // Two start signal will inform the new cfg is available, trigger agu, and inform all extension that a stream is coming
     val readerStart = Input(Bool())
@@ -73,60 +45,73 @@ class XDMADataPath(
     // TCDM request and response signal
     val tcdmReader = new Bundle {
       val req = Vec(
-        readerparam.rwParam.tcdmParam.numChannel,
+        readerParam.rwParam.tcdmParam.numChannel,
         Decoupled(
           new TcdmReq(
-            readerparam.rwParam.tcdmParam.addrWidth,
-            readerparam.rwParam.tcdmParam.dataWidth
+            readerParam.rwParam.tcdmParam.addrWidth,
+            readerParam.rwParam.tcdmParam.dataWidth
           )
         )
       )
       val rsp = Vec(
-        readerparam.rwParam.tcdmParam.numChannel,
+        readerParam.rwParam.tcdmParam.numChannel,
         Flipped(
           Valid(
-            new TcdmRsp(tcdmDataWidth = readerparam.rwParam.tcdmParam.dataWidth)
+            new TcdmRsp(tcdmDataWidth = readerParam.rwParam.tcdmParam.dataWidth)
           )
         )
       )
     }
     val tcdmWriter = new Bundle {
       val req = Vec(
-        writerparam.rwParam.tcdmParam.numChannel,
+        writerParam.rwParam.tcdmParam.numChannel,
         Decoupled(
           new TcdmReq(
-            writerparam.rwParam.tcdmParam.addrWidth,
-            writerparam.rwParam.tcdmParam.dataWidth
+            writerParam.rwParam.tcdmParam.addrWidth,
+            writerParam.rwParam.tcdmParam.dataWidth
           )
         )
       )
     }
 
     // The data for the cluster-level in/out
-    // Cluster-level input <> Writer
-    // Cluster-level output <> Reader
+    // Cluster-level input -> fromRemote <> Writer
+    // Cluster-level output -> toRemote <> Reader
     val remoteXDMAData = new Bundle {
       val fromRemote = Flipped(
         Decoupled(
           UInt(
-            (writerparam.rwParam.tcdmParam.dataWidth * writerparam.rwParam.tcdmParam.numChannel).W
+            (writerParam.rwParam.tcdmParam.dataWidth * writerParam.rwParam.tcdmParam.numChannel).W
           )
         )
       )
+      val fromRemoteAccompaniedCfg = Output(
+        new XDMADataPathCfgIO(
+          axiParam = writerParam.axiParam,
+          crossClusterParam = writerParam.crossClusterParam
+        )
+      )
+
       val toRemote =
         Decoupled(
           UInt(
-            (writerparam.rwParam.tcdmParam.dataWidth * writerparam.rwParam.tcdmParam.numChannel).W
+            (readerParam.rwParam.tcdmParam.dataWidth * readerParam.rwParam.tcdmParam.numChannel).W
           )
         )
+      val toRemoteAccompaniedCfg = Output(
+        new XDMADataPathCfgIO(
+          axiParam = readerParam.axiParam,
+          crossClusterParam = readerParam.crossClusterParam
+        )
+      )
     }
   })
 
   val reader = Module(
-    new Reader(readerparam.rwParam, moduleNamePrefix = clusterName)
+    new Reader(readerParam.rwParam, moduleNamePrefix = clusterName)
   )
   val writer = Module(
-    new Writer(writerparam.rwParam, moduleNamePrefix = clusterName)
+    new Writer(writerParam.rwParam, moduleNamePrefix = clusterName)
   )
 
   // Connect TCDM memory to reader and writer
@@ -151,9 +136,9 @@ class XDMADataPath(
 
   val readerExtensions = Module(
     new DataPathExtensionHost(
-      readerparam.extParam,
+      readerParam.extParam,
       dataWidth =
-        readerparam.rwParam.tcdmParam.dataWidth * readerparam.rwParam.tcdmParam.numChannel,
+        readerParam.rwParam.tcdmParam.dataWidth * readerParam.rwParam.tcdmParam.numChannel,
       headCut = false,
       tailCut = false,
       halfCut = false,
@@ -171,9 +156,9 @@ class XDMADataPath(
 
   val writerExtensions = Module(
     new DataPathExtensionHost(
-      writerparam.extParam,
+      writerParam.extParam,
       dataWidth =
-        writerparam.rwParam.tcdmParam.dataWidth * writerparam.rwParam.tcdmParam.numChannel,
+        writerParam.rwParam.tcdmParam.dataWidth * writerParam.rwParam.tcdmParam.numChannel,
       headCut = false,
       tailCut = false,
       halfCut = false,
@@ -213,23 +198,51 @@ class XDMADataPath(
   readerDemux.io.out(1) <> writerMux.io.in(1)
   readerDemux.io.out(0) <> io.remoteXDMAData.toRemote
   writerMux.io.in(0) <> io.remoteXDMAData.fromRemote
+
+  // Connect the AccompaniedCfg signal
+  io.remoteXDMAData.fromRemoteAccompaniedCfg.convertFromXDMACfgIO(
+    loopBack = io.writerCfg.loopBack,
+    cfg = io.writerCfg,
+    isReaderSide = false
+  )
+  io.remoteXDMAData.toRemoteAccompaniedCfg.convertFromXDMACfgIO(
+    loopBack = io.readerCfg.loopBack,
+    cfg = io.readerCfg,
+    isReaderSide = true
+  )
+
+  io.remoteXDMAData.fromRemoteAccompaniedCfg.readyToTransmit := Mux(
+    io.writerCfg.loopBack,
+    false.B,
+    writer.io.busy
+  )
+
+  io.remoteXDMAData.toRemoteAccompaniedCfg.readyToTransmit := Mux(
+    io.readerCfg.loopBack,
+    false.B,
+    reader.io.busy
+  )
 }
 
 // Below is the class to determine if chisel generate Verilog correctly
 
 object XDMADataPathEmitter extends App {
-  println(
-    getVerilogString(
-      new XDMADataPath(
-        readerparam = new XDMADataPathParam(
-          rwParam = new ReaderWriterParam,
-          extParam = Seq()
-        ),
-        writerparam = new XDMADataPathParam(
-          rwParam = new ReaderWriterParam,
-          extParam = Seq()
-        )
+  emitVerilog(
+    new XDMADataPath(
+      readerParam = new XDMAParam(
+        axiParam = new AXIParam,
+        crossClusterParam = new CrossClusterParam,
+        rwParam = new ReaderWriterParam,
+        extParam = Seq()
+      ),
+      writerParam = new XDMAParam(
+        axiParam = new AXIParam,
+        crossClusterParam = new CrossClusterParam,
+        rwParam = new ReaderWriterParam,
+        extParam = Seq()
       )
-    )
+    ),
+    Array("--target-dir", "generated")
   )
+
 }
