+++ hw/chisel/src/main/scala/snax/xdma/xdmaTop/XDMATop.scala
@@ -10,7 +10,7 @@ import snax.readerWriter.ReaderWriterParam
 import snax.DataPathExtension._
 import snax.xdma.xdmaFrontend._
 import snax.xdma.DesignParams._
-import os.write
+import snax.xdma.io._
 
 import scala.reflect.runtime.currentMirror
 import scala.tools.reflect.ToolBox
@@ -75,9 +75,21 @@ class XDMATopIO(
         )
       )
     )
+    val fromRemoteAccompaniedCfg = Output(
+      new XDMADataPathCfgIO(
+        axiParam = writerParam.axiParam,
+        crossClusterParam = writerParam.crossClusterParam
+      )
+    )
     val toRemote = Decoupled(
       UInt(
         (writerParam.rwParam.tcdmParam.dataWidth * writerParam.rwParam.tcdmParam.numChannel).W
+      )
+    )
+    val toRemoteAccompaniedCfg = Output(
+      new XDMADataPathCfgIO(
+        axiParam = readerParam.axiParam,
+        crossClusterParam = readerParam.crossClusterParam
       )
     )
   }
@@ -112,8 +124,8 @@ class XDMATop(
 
   val xdmaDatapath = Module(
     new XDMADataPath(
-      readerparam = readerParam,
-      writerparam = writerParam,
+      readerParam = readerParam,
+      writerParam = writerParam,
       clusterName = clusterName
     )
   )
@@ -126,8 +138,7 @@ class XDMATop(
   io.tcdmWriter <> xdmaDatapath.io.tcdmWriter
 
   // IO1: Start to connect datapath to axi
-  io.remoteXDMAData.fromRemote <> xdmaDatapath.io.remoteXDMAData.fromRemote
-  io.remoteXDMAData.toRemote <> xdmaDatapath.io.remoteXDMAData.toRemote
+  io.remoteXDMAData <> xdmaDatapath.io.remoteXDMAData
 
   // IO2: Start to coonect ctrl to csr
   io.csrIO <> xdmaCtrl.io.csrIO
